requirements.txt
certifi==2020.6.20
python-dotenv==0.20.0
wincertstore==0.2
